Reworked player boards.

+++ import.sh
@@ -1,3 +1,8 @@
 #! /bin/bash
 input_save="$HOME/.local/share/Tabletop Simulator/Saves/TS_Save_200.json"
-cp "$input_save" 'input.mod.json'
+if [ -f "$input_save" ] ; then
+    cp "$input_save" 'input.mod.json'
+else
+    echo "Boostrapping by creating $input_save"
+    cp 'input.mod.json' "$input_save"
+fi
